ReviewofCode/Thoughts
I noticed that in the code in its present state, deleting objects(e.g. users, groups, etc. ) will only remove them from the table. However it will not remove them from the directories on the server and other locations they might be present. Will have to look into that. This is just to store my thoughts for now.