Updated to determine where the files are located, depending on groups or personal and...
It has been updated to where if upload file button is clicked before anything has been selected, it will pop up an error and reload the page. Also fixes the issue of uploading a file with an apostrophe. It now just removes them and stores them on the server clean.